1.7 Schema Update + ManualProfile

+++ SVP/ManualProfile.cs
@@ -14,6 +14,6 @@ namespace SVP
     
     public partial class ManualProfile : Profile
     {
-        public short ShotCount { get; set; }
+        public int ShotCount { get; set; }
     }
 }
